Force correct placement of function prototypes
The Arduino sketch preprocessor automatically generates function prototypes for any function in a .ino file that doesn't already have a prototype. This system usually works perfectly, but in some rare cases the prototype is inserted at the wrong place in the code. That issue was ocurring in this sketch when compiled with modern versions of the Arduino development software. The solution is to manually add the problematic function prototypes in the correct place.

It's likely that the function prototypes happened to be inserted at the correct location at the time this sketch was written. In the seven years since that time, Arduino has done a major reworking of the prototype generation system. This fixed a lot of common failures of the system, but unfortunately it also created some new ones, which is what happened here.